backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy>=2.0.30
PyMySQL==1.1.0
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy>=3.1.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0
phonenumbers==8.13.25
cryptography==43.0.3
# Python 3.13 compatibility
typing-extensions>=4.8.0
